Move Amount and Description into transaction package
