+++ src/main/java/seedu/ichifund/logic/parser/ParserUtil.java
@@ -9,8 +9,8 @@ import java.util.Set;
 import seedu.ichifund.commons.core.index.Index;
 import seedu.ichifund.commons.util.StringUtil;
 import seedu.ichifund.logic.parser.exceptions.ParseException;
-import seedu.ichifund.model.Amount;
-import seedu.ichifund.model.Description;
+import seedu.ichifund.model.transaction.Amount;
+import seedu.ichifund.model.transaction.Description;
 import seedu.ichifund.model.date.Day;
 import seedu.ichifund.model.date.Month;
 import seedu.ichifund.model.date.Year;
