+++ src/main/java/seedu/ichifund/logic/parser/budget/AddBudgetCommandParser.java
@@ -16,8 +16,8 @@ import seedu.ichifund.logic.parser.Parser;
 import seedu.ichifund.logic.parser.ParserUtil;
 import seedu.ichifund.logic.parser.Prefix;
 import seedu.ichifund.logic.parser.exceptions.ParseException;
-import seedu.ichifund.model.Amount;
-import seedu.ichifund.model.Description;
+import seedu.ichifund.model.transaction.Amount;
+import seedu.ichifund.model.transaction.Description;
 import seedu.ichifund.model.budget.Budget;
 import seedu.ichifund.model.date.Month;
 import seedu.ichifund.model.date.Year;
