+++ src/main/java/seedu/ichifund/storage/JsonAdaptedBudget.java
@@ -4,8 +4,8 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.ichifund.commons.exceptions.IllegalValueException;
-import seedu.ichifund.model.Amount;
-import seedu.ichifund.model.Description;
+import seedu.ichifund.model.transaction.Amount;
+import seedu.ichifund.model.transaction.Description;
 import seedu.ichifund.model.budget.Budget;
 import seedu.ichifund.model.date.Month;
 import seedu.ichifund.model.date.Year;
