+++ src/test/java/seedu/ichifund/testutil/RepeaterBuilder.java
@@ -1,7 +1,7 @@
 package seedu.ichifund.testutil;
 
-import seedu.ichifund.model.Amount;
-import seedu.ichifund.model.Description;
+import seedu.ichifund.model.transaction.Amount;
+import seedu.ichifund.model.transaction.Description;
 import seedu.ichifund.model.repeater.MonthOffset;
 import seedu.ichifund.model.repeater.Repeater;
 import seedu.ichifund.model.transaction.Category;
